split transformations out of dsv loading code for reuse with streaming input
